fix(src-components): renamed files in src/components folder to common naming style
