+++ src/components/pages/AdminBoardDetail/BoardConnectedData/style.scss
