+++ src/components/pages/AdminBoardDetail/BoardHeader/style.scss
