+++ src/components/pages/AdminBoardDetail/BoardPrivacy/style.scss
