+++ src/components/pages/AdminBoardDetail/BoardProvider/style.scss
