+++ src/components/pages/AdminBoardDetail/BoardSecurityInfo/style.scss
