+++ src/components/pages/AdminBoardDetail/BoardTechnicalUserSetup/index.tsx
