+++ src/components/pages/AppUserManagement/AppUserDetailsHeader/style.scss
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-@import '../../../../styles/colors';
+@import '../../../styles/colors';
 
 .app-user-details-header-subtitle {
   text-align: center !important;
