+++ src/components/pages/EdcConnector/AddConnectorOverlay/index.tsx
@@ -29,8 +29,8 @@ import {
   CircleProgress,
   Typography,
 } from '@catena-x/portal-shared-components'
-import ConnectorTypeSelection from './components/ConnectorTypeSelection'
-import ConnectorInsertForm from './components/ConnectorInsertForm'
+import ConnectorTypeSelection from './ConnectorTypeSelection'
+import ConnectorInsertForm from './ConnectorInsertForm'
 import { useForm } from 'react-hook-form'
 import {
   type ConnectorType,
