+++ src/components/pages/Home/AppStoreSection/style.scss
