+++ src/components/pages/Home/BusinessApplicationsSection/index.tsx
