+++ src/components/pages/Home/MenuInfo/index.tsx
