+++ src/components/pages/Home/MenuInfo/style.scss
