+++ src/components/pages/Home/NewsSection/index.tsx
