+++ src/components/pages/Home/NewsSection/style.scss
