+++ src/components/pages/Home/SearchSection/style.scss
