+++ src/components/pages/Home/StageSection/index.tsx
