+++ src/components/shared/basic/ReleaseProcess/components/CommonValidateAndPublish.tsx
@@ -67,7 +67,7 @@ import { useTranslation } from 'react-i18next'
 import { uniqueId } from 'lodash'
 import { PrivacyPolicyType } from 'features/adminBoard/adminBoardApiSlice'
 import { Apartment, Person, LocationOn, Web, Info } from '@mui/icons-material'
-import '../../../../pages/AppDetail/components/AppDetailPrivacy/style.scss'
+import '../../../../pages/AppDetail/AppDetailPrivacy/style.scss'
 import 'components/styles/document.scss'
 
 export interface DefaultValueType {
