+++ src/types/Config.tsx
@@ -55,17 +55,17 @@ import { ACTIONS, HINTS, OVERLAYS, PAGES, ROLES } from './Constants'
 import type { IAction, RestrictedItem, IPage } from './MainTypes'
 import AppUserManagement from 'components/pages/AppUserManagement'
 import IDPManagement from 'components/pages/IDPManagement'
-import AppReleaseProcessForm from 'components/pages/AppReleaseProcess/components'
+import AppReleaseProcessForm from 'components/pages/AppReleaseProcess/AppReleaseProcessForm'
 import CompanyRoles from 'components/pages/CompanyRoles'
 import UseCase from 'components/pages/UseCase'
 import Deactivate from 'components/pages/AppOverview/Deactivate'
 import AdminBoard from 'components/pages/AdminBoard'
 import AdminBoardDetail from 'components/pages/AdminBoardDetail'
 import ServiceReleaseProcess from 'components/pages/ServiceReleaseProcess'
-import ServiceReleaseProcessForm from 'components/pages/ServiceReleaseProcess/components'
+import ServiceReleaseProcessForm from 'components/pages/ServiceReleaseProcess/ServiceReleaseProcessForm'
 import RoleDetails from 'components/pages/RoleDetails'
-import ServiceOverview from 'components/pages/ServiceReleaseProcess/components/ServiceListOverview'
-import ServiceDetails from 'components/pages/ServiceReleaseProcess/components/ServiceDetails'
+import ServiceOverview from 'components/pages/ServiceReleaseProcess/ServiceOverview/ServiceListOverview'
+import ServiceDetails from 'components/pages/ServiceReleaseProcess/ServiceOverview/ServiceDetails'
 import ServiceSubscription from 'components/pages/ServiceSubscription'
 import ServiceAdminBoard from 'components/pages/ServiceAdminBoard'
 import ServiceAdminBoardDetail from 'components/pages/ServiceAdminBoardDetail'
@@ -78,7 +78,7 @@ import ChangeDescription from 'components/pages/AppOverview/ChangeDescription'
 import DataSpace from 'components/pages/DataSpace'
 import AdminCredential from 'components/pages/AdminCredential'
 import AddRoles from 'components/pages/AppOverview/AddRoles'
-import ServiceDeactivate from 'components/pages/ServiceReleaseProcess/components/ServiceDeactivate'
+import ServiceDeactivate from 'components/pages/ServiceReleaseProcess/ServiceOverview/ServiceDeactivate'
 import ChangeDocuments from 'components/pages/AppOverview/ChangeDocuments'
 import CompanyWallet from 'components/pages/CompanyWallet'
 import CompanyCertificates from 'components/pages/CompanyCertificates'
